janus-server/requirements.txt
# Janus-Pro-7B Image Generation Server Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ML/DL
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Janus model (DeepSeek official)
# Install from: https://github.com/deepseek-ai/Janus
# pip install git+https://github.com/deepseek-ai/Janus.git
janus @ git+https://github.com/deepseek-ai/Janus.git

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# HTTP client (for health checks)
httpx>=0.25.0
